Fixing clang compilation error

+++ source/IOHaptics/test/IOBinaryBands.test.cpp
@@ -95,7 +95,7 @@ TEST_CASE("write/read BandHeader on curve") {
     auto byteStreamSize = bitStreamSize % haptics::io::BYTE_SIZE == 0
                               ? bitStreamSize / haptics::io::BYTE_SIZE
                               : (bitStreamSize / haptics::io::BYTE_SIZE) + 1;
-    CHECK(std::filesystem::file_size(filename) == byteStreamSize);
+    CHECK(std::filesystem::file_size(filename) == static_cast<uintmax_t>(byteStreamSize));
   }
 
   SECTION("read band header") {
@@ -296,7 +296,8 @@ TEST_CASE("write/read BandBody on curve") {
     auto byteStreamSize = bitStreamSize % haptics::io::BYTE_SIZE == 0
                               ? bitStreamSize / haptics::io::BYTE_SIZE
                               : (bitStreamSize / haptics::io::BYTE_SIZE) + 1;
-    CHECK(static_cast<int>(std::filesystem::file_size(filename)) == byteStreamSize);
+    CHECK(static_cast<int>(std::filesystem::file_size(filename)) ==
+          static_cast<uintmax_t>(byteStreamSize));
   }
 
   SECTION("read band body") {
